+++ client/app/flux/stores/task-store.js
@@ -6,6 +6,7 @@ import AppDispatcher from '../dispatcher/app-dispatcher';
 import TaskConstants from '../constants/task-constants';
 import Store from './store';
 import Crud from '../../lib/crud';
+import _ from 'lodash';
 import MockTaskData from '../../flux/spec/fixtures/mock-task-data';
 
 class TaskStore extends Store {
@@ -46,23 +47,52 @@ let taskStoreInstance = new TaskStore();
 taskStoreInstance.dispatchToken = AppDispatcher.register(action => {
   switch (action.actionType) {
     case TaskConstants.ADD_TASK:
-      taskStoreInstance.db.post('task', action.data)
-        .then(() => {
-          taskStoreInstance.emitChange();
-        });
+      if (taskStoreInstance.mock) {
+        taskStoreInstance.tasks.push(action.data);
+        taskStoreInstance.emitChange();
+      } else {
+        taskStoreInstance.db.post('task', action.data)
+          .then(() => {
+            taskStoreInstance.emitChange();
+          });
+      }
       break;
+
     case TaskConstants.REMOVE_TASK:
-      taskStoreInstance.db.delete(`task/${action.data}`)
-        .then(() => {
-          taskStoreInstance.emitChange();
+      if (taskStoreInstance.mock) {
+        const taskIndex = _.findIndex(taskStoreInstance.tasks, (task) => {
+          return task.id === action.id;
         });
+
+        if (taskIndex !== -1) {
+          taskStoreInstance.tasks.splice(taskIndex, 1);
+          taskStoreInstance.emitChange();
+        }
+      } else {
+        taskStoreInstance.db.delete(`task/${action.data}`)
+          .then(() => {
+            taskStoreInstance.emitChange();
+          });
+      }
       break;
+
     case TaskConstants.UPDATE_TASK:
-      taskStoreInstance.db.update(`task/${action.id}`, action.data)
-        .then(() => {
-          taskStoreInstance.emitChange();
+      if (taskStoreInstance.mock) {
+        const taskIndex = _.findIndex(taskStoreInstance.tasks, (task) => {
+          return task.id === action.id;
         });
+        if (taskIndex !== -1) {
+          _.assignIn(taskStoreInstance.tasks[taskIndex], action.data);
+          taskStoreInstance.emitChange();
+        }
+      } else {
+        taskStoreInstance.db.update(`task/${action.id}`, action.data)
+          .then(() => {
+            taskStoreInstance.emitChange();
+          });
+      }
       break;
+
     default:
     // no op
   }
