+++ client/app/flux/stores/user-store.js
@@ -5,8 +5,9 @@
 import AppDispatcher from '../dispatcher/app-dispatcher';
 import UserConstants from '../constants/user-constants';
 import Store from './store';
-import MockUsers from '../spec/fixtures/mock-user-data';
 import Crud from '../../lib/crud';
+import MockUserData from '../spec/fixtures/mock-user-data';
+import _ from 'lodash';
 
 class UserStore extends Store {
   constructor(props) {
@@ -22,7 +23,7 @@ class UserStore extends Store {
 
   useMockData() {
     this.mock = true;
-    this.users = MockUsers;
+    this.users = MockUserData;
     this.currentUser = this.users[0];
   }
 
@@ -30,7 +31,7 @@ class UserStore extends Store {
 
     return new Promise((resolve, reject) => {
       if (this.mock) {
-        resolve(MockUsers);
+        resolve(this.users);
       } else {
         this.db.get('task', params)
           .then((data) => {
@@ -53,27 +54,58 @@ let userStoreInstance = new UserStore();
 
 userStoreInstance.dispatchToken = AppDispatcher.register(action => {
   switch (action.actionType) {
-    case UserConstants.ADD_TASK:
-      userStoreInstance.db.post('user', action.data)
-        .then(() => {
-          userStoreInstance.emitChange();
-        });
+    case UserConstants.ADD_USER:
+      if (userStoreInstance.mock) {
+        userStoreInstance.users.push(action.data);
+        userStoreInstance.emitChange();
+      } else {
+        userStoreInstance.db.post('user', action.data)
+          .then(() => {
+            userStoreInstance.emitChange();
+          });
+      }
       break;
-    case UserConstants.REMOVE_TASK:
-      userStoreInstance.db.delete(`user/${action.data}`)
-        .then(() => {
-          userStoreInstance.emitChange();
+
+    case UserConstants.REMOVE_USER:
+      if (userStoreInstance.mock) {
+        const userIndex = _.findIndex(userStoreInstance.users, (user) => {
+          return user.id === action.id;
         });
-      break;
-    case UserConstants.UPDATE_TASK:
-      userStoreInstance.db.update(`user/${action.id}`, action.data)
-        .then(() => {
+
+        if (userIndex !== -1) {
+          userStoreInstance.users.splice(userIndex, 1);
           userStoreInstance.emitChange();
-        });
+        }
+      } else {
+        userStoreInstance.db.delete(`user/${action.data}`)
+          .then(() => {
+            userStoreInstance.emitChange();
+          });
+      }
       break;
+
+    case UserConstants.UPDATE_USER:
+      if (userStoreInstance.mock) {
+        const userIndex = _.findIndex(userStoreInstance.users, (user) => {
+          return user.id === action.id;
+        });
+        if (userIndex !== -1) {
+          _.assignIn(userStoreInstance.users[userIndex], action.data);
+          userStoreInstance.emitChange();
+        }
+      } else {
+        userStoreInstance.db.update(`user/${action.id}`, action.data)
+          .then(() => {
+            userStoreInstance.emitChange();
+          });
+      }
+
+      break;
+
     default:
     // no op
   }
+
 });
 
 export default userStoreInstance;
